+++ src/shared/mod.rs
@@ -1,9 +1,11 @@
+pub use bom::*;
 pub use byte_range::*;
 pub use ctrlc::*;
 pub use hexstring::*;
 pub use recording::*;
 pub(crate) use url::*;
 
+mod bom;
 mod byte_range;
 mod ctrlc;
 mod hexstring;